Refactor admin panel structure and styles
