+++ src/components/adminPanel/AdminLayout.module.css
@@ -1,0 +1,12 @@
+.container {
+    display: flex;
+    gap: 50px;
+    padding: 20px;
+}
+
+@media (max-width: 1024px) {
+    .container {
+        flex-direction: column;
+        gap: 20px;
+    }
+}
