+++ src/components/adminPanel/ContentArea/style.module.css
