+++ src/components/adminPanel/FullPost/index.jsx
