+++ src/components/adminPanel/FullPost/style.module.css
