+++ src/components/adminPanel/Menu/images/Switch.png
