+++ src/components/adminPanel/Password/edit.jsx
@@ -21,7 +21,7 @@ function EditPassword() {
   };
 
   return (
-    <div>
+    <div className={style.container}>
       <Typography
         sx={{
           fontFamily: "SourceCodePro-SemiBold",
