+++ src/components/adminPanel/Post/edit.module.css
