+++ src/components/adminPanel/Skeleton/skeleton.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Skeleton from '@mui/material/Skeleton';
-import s from '../post/style.module.css'
+import s from '../Post/style.module.css'
 
 export default function PostLoading() {
   return (
@@ -14,5 +14,4 @@ export default function PostLoading() {
             <button className={s.btnEdit}>Изменить</button>
         </div>
       </div>
-  );
-}
+  );}
