--- src/components/adminPannel/password/style.module.css
@@ -1,4 +1,0 @@
-.input {
-  padding: 10px;
-  margin-bottom: 20px;
-}
